Fix
  Line   tests\Backend\Snowflake\TimeTravelTest.php
 ------ ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
         Ignored error pattern #^Parameter \#2 \$timestamp of function date expects int, int\|false given\.$# in path C:\dev\keboola\storage-api-php-client\tests\Backend\Snowflake\TimeTravelTest.php was
         not matched in reported errors.
  54     Parameter #2 $timestamp of function date expects int|null, int|false given.
  105    Parameter #2 $timestamp of function date expects int|null, int|false given.

+++ tests/Backend/Snowflake/TimeTravelTest.php
@@ -51,7 +51,7 @@ class TimeTravelTest extends StorageApiTestCase
 
         $updatedTable = $this->_client->getTable($sourceTableId);
 
-        $newTableName = "new-table-name_" . date('Ymd_His', strtotime($timestamp));
+        $newTableName = "new-table-name_" . date('Ymd_His', (int) strtotime($timestamp));
 
         $replicaTableId = $this->_client->createTableFromSourceTableAtTimestamp(
             $this->getTestBucketId(self::STAGE_OUT),
@@ -102,7 +102,7 @@ class TimeTravelTest extends StorageApiTestCase
 
         $updatedTable = $this->_client->getTable($sourceTableId);
 
-        $newTableName = "new-table-name_" . date('Ymd_His', strtotime($timestamp));
+        $newTableName = "new-table-name_" . date('Ymd_His', (int) strtotime($timestamp));
 
         $replicaTableId = $this->_client->createTableFromSourceTableAtTimestamp(
             $this->getTestBucketId(self::STAGE_OUT),
